Moved getMaps to Slide methods instead of Slider
